[level 2] Title: n^2 배열 자르기, Time: 4.34 ms, Memory: 109 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -7,9 +7,9 @@ class Solution {
         for (int i=0;i<dif;i++){
             
             long k= i+left;
-            long x = k/n;
-            long y = k%n;
-            answer[i] = (int)Math.max(x,y)+1;
+            int x = (int)(k/n);
+            int y = (int)(k%n);
+            answer[i] = Math.max(x,y)+1;
             
         }
